플레이어 체력 안깎이는버그 수정

+++ Client_Framework/CEntityState.h
@@ -13,8 +13,8 @@ protected:
 	
 
 public:
-	int mHealth{};
-	int mMaxHealth{};
+	float mHealth{};
+	float mMaxHealth{};
 	bool mIsHit = false;
 	bool mIsDead = false;
 	bool mIsSpawning = false;
